[vim] Prevent two spaces after `.` when joining with `J`
In normal mode, `J` merges the current line with the line below. If the
current line ends with a `.` denoting the end of a sentence, vim appends
an extra space after the `.` which results in `.  ` instead of the
preferred `. `

This configuration joins sentences the latter way described, with
a single space